requirements.txt
black==21.7b0
Django==3.2.5
djangorestframework==3.12.4
pytest-django==4.4.0
